Remove support for Snapshot alpha and beta APIs
